refactor(desktop): move chat components and utils to feature domain
