+++ apps/desktop/src/__tests__/ParallelProgressPill.test.tsx
@@ -2,7 +2,7 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 import { cleanup, fireEvent, render, screen } from '@testing-library/react';
 import type { AgentStatus, ProviderRunId } from '@kay-am/types';
-import { ParallelProgressPill } from '../components/chat/ParallelProgressPill';
+import { ParallelProgressPill } from '../features/chat/components/ParallelProgressPill';
 
 function rid(s: string): ProviderRunId {
   return s as ProviderRunId;
