+++ apps/desktop/src/features/chat/components/ParallelProgressPill/index.tsx
