+++ apps/desktop/src/features/chat/components/RoutingIndicator/index.tsx
@@ -5,8 +5,8 @@ import type {
   SessionProviderPreference,
   TurnProviderOverride,
 } from '@kay-am/types';
-import { resolveProviderForTurn } from '../../features/providers/routing';
-import { SESSION_FEATURES } from '../../shared/lib/features';
+import { resolveProviderForTurn } from '../../../../features/providers/routing';
+import { SESSION_FEATURES } from '../../../../shared/lib/features';
 
 interface RoutingIndicatorProps {
   readonly sessionPreference: SessionProviderPreference;
