+++ apps/desktop/src/features/chat/components/SlashCommandPopover/index.tsx
