+++ apps/desktop/src/features/chat/utils/format-card-time.ts
